Remove typescript e corrige alguns bugs
