+++ package.json
@@ -9,20 +9,16 @@
   "dependencies": {
     "@grammyjs/conversations": "^1.1.2",
     "@grammyjs/hydrate": "^1.3.1",
-    "@grammyjs/menu": "^1.2.1",
+    "@grammyjs/menu": "^1.2.0",
     "@grammyjs/parse-mode": "github:grammyjs/parse-mode",
-    "@types/node-persist": "^3.1.4",
     "dotenv": "^16.3.1",
     "grammy": "^1.18.1",
+    "i": "^0.3.7",
     "node-persist": "^3.1.3",
+    "npm": "^10.1.0",
     "parse-mode": "^0.1.4"
   },
-  "devDependencies": {
-    "ts-node": "^10.9.1",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^4.7.4"
-  },
   "scripts": {
-    "start": "ts-node-dev --respawn --transpile-only ./src/index.ts"
+    "start": "node ./src/index.js"
   }
 }
