+++ src/handlers/cube/removeCube.js
@@ -1,0 +1,3 @@
+export async function removeCube(conversation, ctx) {
+  ctx.reply("essa função remove itens no cubo");
+}
