--- src/handlers/cube/removeCube.ts
@@ -1,6 +1,0 @@
-
-export async function removeCube(conversation: MyConversation, ctx: MyContext){
-  ctx.reply("essa função remove itens no cubo");
-}
-
-
