+++ src/handlers/modifyItem/index.js
@@ -1,0 +1,5 @@
+const modifyItem = () => {};
+
+module.exports = {
+  modifyItem,
+};
