--- src/interfaces/character.ts
@@ -1,9 +1,0 @@
-import { Item } from "./item";
-
-export interface Character {
-  id: string;
-  name: string;
-  level: number;
-  vocation: string;
-  items: Item[];
-}
