--- src/interfaces/item.ts
@@ -1,6 +1,0 @@
-export interface Item {
-  name: string;
-  weight: number;
-  quantity: number;
-  description: string;
-}
